tg/requirements.txt
python-telegram-bot>=20.7
Flask>=2.0
requests>=2.25